Fix UI bug with content overflow
A long list of names or bills was spilling over into the header and
footer. This moves the footer outside the main container and rewrites
the styling for navbar, main container, and footer so that all content
is visible.

Since long lists now have the potential to push the footer out of view,
the essential footer buttons now need to be brought up to the navbar.